Refactoring
Refactoring to clean up fit_fcn
Also changing imports slightly
