+++ src/optimize/CheKiPEUQ_from_Frhodo.py
@@ -1,31 +1,23 @@
 import numpy as np
-import sys; sys.path.append('../../'); import os
-sys.path.append(os.path.join(os.path.dirname(__file__))) #This is so that CheKiPEUQ_local can be found properly. Otherwise there are problems importing CKPQ.parameter_estimation
+from copy import deepcopy
 
-# try:
-    # import CheKiPEUQ as CKPQ
-    # print("line 5 it is importing CheKiPEUQ!")
-# except:
-    # import os #The below lines are to allow CheKiPEUQ_local to be called regardless of user's working directory.
-    # lenOfFileName = len(os.path.basename(__file__)) #This is the name of **this** file.
-    # absPathWithoutFileName = os.path.abspath(__file__)[0:-1*lenOfFileName]
-    # sys.path.append(absPathWithoutFileName)
-    # import optimize.CheKiPEUQ_local as CKPQ #might need to put Frhodo.CheKiPEUQ_local or something like that.
-    # # compare to C:\Users\fvs\Documents\GitHub\CheKiPEUQ\CheKiPEUQ\InverseProblem.py
-    # print("line 14 it is in the except statement of importing CheKiPEUQ!")
+try:
+    import CheKiPEUQ as CKPQ
+except:
+    #import pathlib
+    #sys.path.append(pathlib.Path(__file__).parent.absolute()) # add directory of **this** file to path
+    import optimize.CheKiPEUQ_local as CKPQ 
     
-lenOfFileName = len(os.path.basename(__file__)) #This is the name of **this** file.
-absPathWithoutFileName = os.path.abspath(__file__)[0:-1*lenOfFileName]
-sys.path.append(absPathWithoutFileName)
-import optimize.CheKiPEUQ_local as CKPQ #might need to put Frhodo.CheKiPEUQ_local or something like that.
-    
+try:
+    import CheKiPEUQ.UserInput as UserInput
+except:
+    import optimize.CheKiPEUQ_local.UserInput as UserInput
+
 try:
     import CiteSoft
 except:
-    import os #The below lines are to allow CiteSoftLocal to be called regardless of user's working directory.
-    lenOfFileName = len(os.path.basename(__file__)) #This is the name of **this** file.
-    absPathWithoutFileName = os.path.abspath(__file__)[0:-1*lenOfFileName]
-    sys.path.append(absPathWithoutFileName)
+    #import pathlib #The below lines are to allow CiteSoftLocal to be called regardless of user's working directory.
+    #sys.path.append(pathlib.Path(__file__).parent.absolute()) # add directory of **this** file to path
     import CiteSoftLocal as CiteSoft
 
 #get things ready for CiteSoft entry...
@@ -48,11 +40,7 @@ def load_into_CheKiPUEQ(simulation_function, observed_data, pars_initial_guess =
     #sigma_multiple is how many sigma the bounds are equal to (relative to mean).
     #pars_uncertainty_distribution allows 3 choices: automatic, gaussian, uniform.  Automatic gives 'uniform' to the rate constants and 'gaussian' to the rate_constant_parameters.
     #num_rate_constants_and_rate_constant_parameters  allows the 'automatic setting of pars_uncertainty_distribution to assign distribution types based on the par index. If not supplied, everything is assumed to be a rate_constant.
-    try:
-        import CheKiPEUQ.UserInput as UserInput
-        print("line 44, somehow imported CheKiPEUQ UserInput!!!")
-    except:
-        import optimize.CheKiPEUQ_local.UserInput as UserInput
+    
     #TODO: put a "clear UserInput" type call here to UnitTesterSG_local
     
     if len(num_rate_constants_and_rate_constant_parameters) == 0: num_rate_constants_and_rate_constant_parameters = [len(pars_initial_guess), 0]
@@ -163,3 +151,115 @@ def get_consolidated_parameters_arrays(rate_constants_values, rate_constants_low
 def remove_unbounded_values(array_to_truncate, unbounded_indices):
     truncated_array = np.delete(array_to_truncate, unbounded_indices, axis=0)
     return truncated_array
+
+class CheKiPEUQ:
+    def __init__(self, input_dict):
+        #Step 1 of Bayesian:  Prepare any variables that need to be passed in for Bayesian PE_object creation.
+        #Step 2 of Bayesian:  populate Bayesian_dict with any variables and uncertainties needed.
+        self.Bayesian_dict = {}
+        self.Bayesian_dict['pars_uncertainty_distribution'] = input_dict['opt_settings']['bayes_dist_type']  #options can be 'Auto', 'Gaussian' or 'Uniform'. 
+        self.Bayesian_dict['rate_constants_initial_guess'] = deepcopy(input_dict['x0'])
+        self.Bayesian_dict['rate_constants_lower_bnds'] = deepcopy(input_dict['bounds']['lower'])
+        self.Bayesian_dict['rate_constants_upper_bnds'] = deepcopy(input_dict['bounds']['upper'])
+        num_rate_consants = len(self.Bayesian_dict['rate_constants_initial_guess'])
+        self.Bayesian_dict['rate_constants_bnds_exist'] =  np.array(np.ones((num_rate_consants, 2)), dtype = bool) #From Jan 2021, we are setting [True True] for each rate_constant.
+
+        self.Bayesian_dict['rate_constants_parameters_changing'] = deepcopy(input_dict['coef_opt'])
+        self.Bayesian_dict['rate_constants_parameters_initial_guess'] = []
+        self.Bayesian_dict['rate_constants_parameters_lower_bnds'] = []
+        self.Bayesian_dict['rate_constants_parameters_upper_bnds'] = []
+        self.Bayesian_dict['rate_constants_parameters_bnds_exist'] = []
+        for rxn_coef in input_dict['rxn_coef_opt']:
+            self.Bayesian_dict['rate_constants_parameters_initial_guess'].append(deepcopy(rxn_coef['coef_x0'])) 
+            self.Bayesian_dict['rate_constants_parameters_lower_bnds'].append(deepcopy(rxn_coef['coef_bnds']['lower']))
+            self.Bayesian_dict['rate_constants_parameters_upper_bnds'].append(deepcopy(rxn_coef['coef_bnds']['upper']))
+            self.Bayesian_dict['rate_constants_parameters_bnds_exist'].extend(list(deepcopy(rxn_coef['coef_bnds']['exist']))) #we can't use append because this is a list of lists/array, and we want parallel construction.
+        self.Bayesian_dict['rate_constants_parameters_initial_guess'] = np.array(self.Bayesian_dict['rate_constants_parameters_initial_guess']).flatten()
+        self.Bayesian_dict['rate_constants_parameters_lower_bnds'] = np.array(self.Bayesian_dict['rate_constants_parameters_lower_bnds']).flatten()
+        self.Bayesian_dict['rate_constants_parameters_upper_bnds'] = np.array(self.Bayesian_dict['rate_constants_parameters_upper_bnds']).flatten()
+        #self.Bayesian_dict['rate_constants_parameters_bnds_exist'] does not get flattened because it is a list of list/arrays (so we don't want it flattened).
+
+        Bayesian_dict = self.Bayesian_dict            
+        #concatenate all of the initial guesses and bounds. 
+        Bayesian_dict['pars_initial_guess'], Bayesian_dict['pars_lower_bnds'],Bayesian_dict['pars_upper_bnds'], Bayesian_dict['pars_bnds_exist'], Bayesian_dict['unbounded_indices'] = get_consolidated_parameters_arrays( 
+            Bayesian_dict['rate_constants_initial_guess'],
+            Bayesian_dict['rate_constants_lower_bnds'],
+            Bayesian_dict['rate_constants_upper_bnds'],                
+            Bayesian_dict['rate_constants_bnds_exist'],
+            Bayesian_dict['rate_constants_parameters_initial_guess'],
+            Bayesian_dict['rate_constants_parameters_lower_bnds'],
+            Bayesian_dict['rate_constants_parameters_upper_bnds'],
+            Bayesian_dict['rate_constants_parameters_bnds_exist'],
+            return_unbounded_indices=True
+            )
+        #remove the unbounded values.
+        Bayesian_dict['pars_initial_guess_truncated'] = remove_unbounded_values(Bayesian_dict['pars_initial_guess'], Bayesian_dict['unbounded_indices'] )
+        Bayesian_dict['pars_lower_bnds_truncated'] = remove_unbounded_values( Bayesian_dict['pars_lower_bnds'], Bayesian_dict['unbounded_indices'] )
+        Bayesian_dict['pars_upper_bnds_truncated'] = remove_unbounded_values( Bayesian_dict['pars_upper_bnds'], Bayesian_dict['unbounded_indices'] )
+        Bayesian_dict['pars_bnds_exist_truncated'] = remove_unbounded_values( Bayesian_dict['pars_bnds_exist'], Bayesian_dict['unbounded_indices'] )
+
+    def evaluate(self, CheKiPEUQ_eval_dict):
+        Bayesian_dict = self.Bayesian_dict                
+        Bayesian_dict['rate_constants_current_guess'] = deepcopy(CheKiPEUQ_eval_dict['log_opt_rates'])
+        Bayesian_dict['rate_constants_parameters_current_guess'] = deepcopy(CheKiPEUQ_eval_dict['x'])
+        Bayesian_dict['pars_current_guess'] = np.concatenate( (Bayesian_dict['rate_constants_current_guess'], Bayesian_dict['rate_constants_parameters_current_guess'] ) )
+        Bayesian_dict['last_obs_sim_interp'] = np.concatenate(CheKiPEUQ_eval_dict['output_dict']['obs_sim_interp'], axis=0)
+        Bayesian_dict['observed_data'] = np.concatenate(CheKiPEUQ_eval_dict['output_dict']['obs_exp'], axis=0)
+        Bayesian_dict['observed_data_lower_bounds'] = []
+        Bayesian_dict['observed_data_upper_bounds'] = []
+        Bayesian_dict['weights_data'] = CheKiPEUQ_eval_dict['bayesian_weights']
+
+        def get_last_obs_sim_interp(varying_rate_vals): #A. Savara added this. It needs to be here.
+            try:
+                last_obs_sim_interp = Bayesian_dict['last_obs_sim_interp']
+                last_obs_sim_interp = np.array(last_obs_sim_interp).T
+            except:
+                print("this is line 422! There may be an error occurring!")
+                last_obs_sim_interp = None
+            return last_obs_sim_interp
+            
+        Bayesian_dict['simulation_function'] = get_last_obs_sim_interp #using wrapper that just returns the last_obs_sim_interp
+           
+        # Step 3 of Bayesian:  create a CheKiPEUQ_PE_Object (this is a class object)
+        # NOTE: normally, the Bayesian object would be created earlier. However, we are using a non-standard application
+        # where the observed_data uncertainties might change with each simulation.
+        # To gain efficiency, we could cheat and se the feature get_responses_simulation_uncertainties function of CheKiPEUQ, 
+        # or could create a new get_responses_observed_uncertainties function in CheKiPEUQ
+        # for now we will just create a new PE_object each time.
+        if np.shape(Bayesian_dict['weights_data']) != np.shape(Bayesian_dict['observed_data']):
+            raise Exception('CheKiPEUQ Error: Shape of weights does not match observed data')
+            
+        CheKiPEUQ_PE_object = load_into_CheKiPUEQ(
+            simulation_function=    Bayesian_dict['simulation_function'],
+            observed_data=          Bayesian_dict['observed_data'],
+            pars_initial_guess =    Bayesian_dict['pars_initial_guess_truncated'],
+            pars_lower_bnds =       Bayesian_dict['pars_lower_bnds_truncated'],   
+            pars_upper_bnds =       Bayesian_dict['pars_upper_bnds_truncated'],   
+            pars_bnds_exist =       Bayesian_dict['pars_bnds_exist_truncated'],
+            observed_data_lower_bounds= Bayesian_dict['observed_data_lower_bounds'],
+            observed_data_upper_bounds= Bayesian_dict['observed_data_upper_bounds'],
+            weights_data=               Bayesian_dict['weights_data'],
+            pars_uncertainty_distribution=  Bayesian_dict['pars_uncertainty_distribution'],
+            num_rate_constants_and_rate_constant_parameters = [len(Bayesian_dict['rate_constants_initial_guess']), len(Bayesian_dict['rate_constants_parameters_initial_guess'])]
+            ) #this is assigned in the "__init__" function up above.
+            
+        # Step 4 of Bayesian:  call a function to get the posterior density which will be used as the objective function.
+        # We need to provide the current values of the varying_rate_vals to feed into the function.
+            
+        #varying_rate_vals = np.array(output_dict['shock']['rate_val'])[list(varying_rate_vals_indices)] #when extracting a list of multiple indices, instead of array[index] one use array[[indices]]
+        Bayesian_dict['pars_current_guess_truncated'] = remove_unbounded_values(Bayesian_dict['pars_current_guess'], Bayesian_dict['unbounded_indices'] ) 
+            
+        log_posterior_density = get_log_posterior_density(CheKiPEUQ_PE_object, Bayesian_dict['pars_current_guess_truncated'])
+        neg_log_posterior_density = -1*log_posterior_density # need neg_logP because minimizing.
+            
+        # Step 5 of Bayesian:  return the objective function and any other metrics desired.
+        #obj_fcn = neg_log_posterior_density 
+        if CheKiPEUQ_eval_dict['iteration_num'] == 0 and 'obj_fcn_initial' not in Bayesian_dict:
+            Bayesian_dict['obj_fcn_initial'] = neg_log_posterior_density 
+            obj_fcn = 0.0
+        elif neg_log_posterior_density == np.inf:
+            obj_fcn = np.inf
+        else:
+            obj_fcn = (neg_log_posterior_density - Bayesian_dict['obj_fcn_initial'])/np.abs(Bayesian_dict['obj_fcn_initial'])*100
+
+        return obj_fcn
