+++ src/optimize/CheKiPEUQ_local/__init__.py
@@ -2,7 +2,7 @@
 #Except instead of putting things here directly in __init__, we'll impor them so they are accessible by importing the module.
 import sys; import os
 #sys.path.append(os.path.join(os.path.dirname(__file__), '../')) #if the directory one above is added to sys.path, then InverseProblem can be found for importing from.
-from CheKiPEUQ_local.InverseProblem import *
+from .InverseProblem import *
 #from CheKiPEUQ.plotting_functions import *
 #from CheKiPEUQ.mumpce import *
-from InverseProblem import parameter_estimation
+#from InverseProblem import parameter_estimation
